// frontend/src/types/model.ts
// Had to add these types.
//Would be nice to get them generated by Prisma for single source of truth but had way too much extra boilerplate going that route.
// Would likely build custom typegen for this project.

/**
 * Model Question
 */
export type Question = {
  id: number;
  title: string;
  body: string;
  creation: number;
  score: number;
  userId: number;
  user: User;
  answers: Answer[];
  comments: Comment[];
};

/**
 * Model Answer
 */
export type Answer = {
  id: number;
  body: string;
  creation: number;
  score: number;
  accepted: boolean;
  questionId: number;
  userId: number;
  user: User;
  comments: Comment[];
};

/**
 * Model Comment
 */
export type Comment = {
  id: number;
  body: string;
  userId: number;
  questionId: number | null;
  answerId: number | null;
  user: User;
};

/**
 * Model User
 */
export type User = {
  id: number;
  name: string;
  questions: Question[];
  answers: Answer[];
  comments: Comment[];
};